// webapp/src/shared/generated/models/Game.ts
/* tslint:disable */
/* eslint-disable */
/**
 * transcendence-app
 * The transcendence-app API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Game
 */
export interface Game {
    /**
     * 
     * @type {string}
     * @memberof Game
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Game
     */
    playerOneUsername: string;
    /**
     * 
     * @type {string}
     * @memberof Game
     */
    playerTwoUsername: string;
    /**
     * 
     * @type {Date}
     * @memberof Game
     */
    createdAt: Date;
    /**
     * 
     * @type {number}
     * @memberof Game
     */
    gameDurationInSeconds: number;
    /**
     * 
     * @type {number}
     * @memberof Game
     */
    playerOneScore: number;
    /**
     * 
     * @type {number}
     * @memberof Game
     */
    playerTwoScore: number;
    /**
     * 
     * @type {string}
     * @memberof Game
     */
    gameMode: GameGameModeEnum;
}


/**
 * @export
 */
export const GameGameModeEnum = {
    Training: 'training',
    Classic: 'classic',
    ShortPaddle: 'shortPaddle',
    MysteryZone: 'mysteryZone',
    Unknown: 'unknown'
} as const;
export type GameGameModeEnum = typeof GameGameModeEnum[keyof typeof GameGameModeEnum];


/**
 * Check if a given object implements the Game interface.
 */
export function instanceOfGame(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "playerOneUsername" in value;
    isInstance = isInstance && "playerTwoUsername" in value;
    isInstance = isInstance && "createdAt" in value;
    isInstance = isInstance && "gameDurationInSeconds" in value;
    isInstance = isInstance && "playerOneScore" in value;
    isInstance = isInstance && "playerTwoScore" in value;
    isInstance = isInstance && "gameMode" in value;

    return isInstance;
}

export function GameFromJSON(json: any): Game {
    return GameFromJSONTyped(json, false);
}

export function GameFromJSONTyped(json: any, ignoreDiscriminator: boolean): Game {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'playerOneUsername': json['playerOneUsername'],
        'playerTwoUsername': json['playerTwoUsername'],
        'createdAt': (new Date(json['createdAt'])),
        'gameDurationInSeconds': json['gameDurationInSeconds'],
        'playerOneScore': json['playerOneScore'],
        'playerTwoScore': json['playerTwoScore'],
        'gameMode': json['gameMode'],
    };
}

export function GameToJSON(value?: Game | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'playerOneUsername': value.playerOneUsername,
        'playerTwoUsername': value.playerTwoUsername,
        'createdAt': (value.createdAt.toISOString()),
        'gameDurationInSeconds': value.gameDurationInSeconds,
        'playerOneScore': value.playerOneScore,
        'playerTwoScore': value.playerTwoScore,
        'gameMode': value.gameMode,
    };
}