// webapp/src/shared/generated/models/UserResponseDto.ts
export interface UserResponseDto {
    /**
     * 
     * @type {string}
     * @memberof UserResponseDto
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponseDto
     */
    username: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponseDto
     */
    email: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponseDto
     */
    fullName: string;
    /**
     * 
     * @type {string}
     * @memberof UserResponseDto
     */
    avatarId: string;
    /**
     * 
     * @type {number}
     * @memberof UserResponseDto
     */
    avatarX: number;
    /**
     * 
     * @type {number}
     * @memberof UserResponseDto
     */
    avatarY: number;
    /**
     * 
     * @type {Date}
     * @memberof UserResponseDto
     */
    createdAt: Date;
    /**
     * 
     * @type {boolean}
     * @memberof UserResponseDto
     */
    isTwoFactorAuthenticationEnabled: boolean;
    /**
     * 
     * @type {number}
     * @memberof UserResponseDto
     */
    level: number;
    /**
     * 
     * @type {UserResponseDtoBlockRelation}
     * @memberof UserResponseDto
     */
    blockRelation: UserResponseDtoBlockRelation | null;
    /**
     * 
     * @type {boolean}
     * @memberof UserResponseDto
     */
    isFriend: boolean | null;
}